backend/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.20
httpx>=0.24
pydantic>=2.0
python-dotenv>=1.0
openai>=1.0
